requirements.txt
flask==2.2.2
mysql-connector-python==8.0.33
